third_party/blink/web_tests/platform/generic/fast/forms/file/file-input-capture-expected.txt

Tests the capture attribute of <input type='file'>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL 'capture' in input should be true. Was false.
FAIL input.capture should be false (of type boolean). Was undefined (of type undefined).
PASS input.hasAttribute('capture') is false
FAIL input.capture should be false (of type boolean). Was undefined (of type undefined).
PASS input.hasAttribute('capture') is false
FAIL input.capture should be true (of type boolean). Was undefined (of type undefined).
PASS input.hasAttribute('capture') is true
FAIL input.capture should be false (of type boolean). Was undefined (of type undefined).
PASS input.hasAttribute('capture') is false
FAIL input.capture should be true (of type boolean). Was undefined (of type undefined).
PASS input.hasAttribute('capture') is true
PASS input.capture is false
FAIL input.hasAttribute('capture') should be false. Was true.
PASS input.capture is true
PASS input.hasAttribute('capture') is true
PASS successfullyParsed is true

TEST COMPLETE
